refactor(models): make the declarative conversion chain the only path
Each model defines `conversion_chain(config)` in its `converting_<model>.py`;
the `PreTrainedModelPrimeRL` base has a single set of `convert_to_hf` /
`convert_to_prime` / `convert_layer_*` implementations that play the chain
forward and backward. Because every op is present-guarded, the same chain
works over a full state dict, one layer's keys, or a local shard.

Removes the now-redundant scaffolding:
- per-model imperative `convert_hf_to_tt` / `convert_tt_to_hf` functions
- the standalone `conversion_chain.py` builders and the `conversion_chains` registry
- the mock equivalence tests (superseded by the 20-step math KL-mismatch
  validation in PR #2797) and the obsolete classmethod roundtrip tests

Net -2.2k lines. Validated end-to-end: qwen3_moe/qwen3_5_moe/nemotron_h/glm4_moe
mean KL mismatch < 0.0015 over 20 steps on math.
