+++ src/prime_rl/trainer/models/glm4_moe/converting_glm4_moe.py
@@ -1,151 +1,67 @@
-import torch
-from torch import Tensor
+"""HF<->prime weight conversion for GLM-4 MoE, as a declarative op chain.
+
+GLM-4 MoE specifics:
+
+* Router: HF ``mlp.gate.weight`` -> prime ``mlp.router.gate.weight``.
+* Expert bias: HF ``mlp.gate.e_score_correction_bias`` -> prime
+  ``mlp.expert_bias``.
+* Routed experts: HF per-expert ``mlp.experts.{e}.{gate,down,up}_proj.weight``
+  stack into prime ``mlp.experts.{w1,w2,w3}`` along dim 0; the fused
+  transformers-v5 ``mlp.experts.gate_up_proj`` / ``down_proj`` layout is also
+  accepted on the way in (split along dim 1).
+* Shared experts: HF ``mlp.shared_experts.{gate,down,up}_proj.weight`` map to
+  prime ``mlp.shared_expert.{w1,w2,w3}`` (no ``.weight`` suffix on prime; this
+  is a plain rename forward). On the way back to HF, GLM strips a leading
+  singleton dim from each shared-expert tensor when present
+  (:class:`~prime_rl.trainer.models.conversion_ops.SqueezeLeading`,
+  backward-only).
+* Prime-only runtime buffer ``mlp.tokens_per_expert`` is dropped on the way
+  back to HF.
+
+The per-layer MoE op builder (:func:`glm_moe_layer_ops`) is shared verbatim
+with GLM-MoE-DSA, whose attention/MLA weights pass through untouched.
+"""
+
+from __future__ import annotations
+
+from prime_rl.trainer.models.conversion_ops import (
+    GATE_DOWN_UP,
+    ConvOp,
+    Drop,
+    Rename,
+    SqueezeLeading,
+    routed_experts_op,
+)
 
 
-def get_max_layer_num(state_dict: dict[str, Tensor]) -> int:
-    """Get the maximum number of layers in the model."""
-    return max(int(i.split(".")[2]) for i in state_dict.keys() if "model.layers." in i) + 1
+def glm_moe_layer_ops(layer_idx: int) -> list[ConvOp]:
+    """The hf<->tt MoE ops for one GLM-style layer.
 
-
-def convert_hf_layer_to_tt(state_dict: dict[str, Tensor], layer_idx: int):
-    i = layer_idx
-
-    # Check if this is a MoE layer by looking for gate weight
-    if f"model.layers.{i}.mlp.gate.weight" not in state_dict:
-        return
-
-    state_dict[f"model.layers.{i}.mlp.router.gate.weight"] = state_dict[f"model.layers.{i}.mlp.gate.weight"]
-    del state_dict[f"model.layers.{i}.mlp.gate.weight"]
-
-    # Check if this is the new fused format (transformers 5.0+) or old per-expert format
-    if f"model.layers.{i}.mlp.experts.gate_up_proj" in state_dict:
-        # New fused format: gate_up_proj has shape (num_experts, 2*moe_dim, dim)
-        gate_up_proj = state_dict[f"model.layers.{i}.mlp.experts.gate_up_proj"]
-        down_proj = state_dict[f"model.layers.{i}.mlp.experts.down_proj"]
-
-        num_experts, fused_dim, dim = gate_up_proj.shape
-        moe_dim = fused_dim // 2
-
-        # Split gate_up_proj into w1 (gate) and w3 (up)
-        w1 = gate_up_proj[:, :moe_dim, :]  # Gate: (num_experts, moe_dim, dim)
-        w3 = gate_up_proj[:, moe_dim:, :]  # Up: (num_experts, moe_dim, dim)
-        w2 = down_proj  # Down: (num_experts, dim, moe_dim)
-
-        del state_dict[f"model.layers.{i}.mlp.experts.gate_up_proj"]
-        del state_dict[f"model.layers.{i}.mlp.experts.down_proj"]
-    else:
-        # Old per-expert format
-        num_experts = len([j for j in state_dict.keys() if f"model.layers.{i}.mlp.experts" in j]) // 3
-        if num_experts == 0:
-            return
-
-        dim, moe_dim = state_dict[f"model.layers.{i}.mlp.experts.0.down_proj.weight"].shape
-        w1 = torch.empty(
-            (num_experts, moe_dim, dim), dtype=state_dict[f"model.layers.{i}.mlp.experts.1.down_proj.weight"].dtype
-        )  # Gate
-        w2 = torch.empty(
-            (num_experts, dim, moe_dim), dtype=state_dict[f"model.layers.{i}.mlp.experts.1.down_proj.weight"].dtype
-        )  # Down
-        w3 = torch.empty(
-            (num_experts, moe_dim, dim), dtype=state_dict[f"model.layers.{i}.mlp.experts.1.down_proj.weight"].dtype
-        )  # Up
-        for j in range(num_experts):
-            w1[j].copy_(state_dict[f"model.layers.{i}.mlp.experts.{j}.gate_proj.weight"])
-            w2[j].copy_(state_dict[f"model.layers.{i}.mlp.experts.{j}.down_proj.weight"])
-            w3[j].copy_(state_dict[f"model.layers.{i}.mlp.experts.{j}.up_proj.weight"])
-
-            del state_dict[f"model.layers.{i}.mlp.experts.{j}.gate_proj.weight"]
-            del state_dict[f"model.layers.{i}.mlp.experts.{j}.down_proj.weight"]
-            del state_dict[f"model.layers.{i}.mlp.experts.{j}.up_proj.weight"]
-
-    state_dict[f"model.layers.{i}.mlp.experts.w1"] = w1
-    state_dict[f"model.layers.{i}.mlp.experts.w2"] = w2
-    state_dict[f"model.layers.{i}.mlp.experts.w3"] = w3
-
-    state_dict[f"model.layers.{i}.mlp.shared_expert.w1"] = state_dict[
-        f"model.layers.{i}.mlp.shared_experts.gate_proj.weight"
+    Shared by GLM-4 MoE and GLM-MoE-DSA (their hf<->tt MoE conversion is
+    identical; DSA's attention/MLA weights pass through untouched)."""
+    p = f"model.layers.{layer_idx}.mlp"
+    ops: list[ConvOp] = [
+        Rename(f"{p}.gate.weight", f"{p}.router.gate.weight"),
+        Rename(f"{p}.gate.e_score_correction_bias", f"{p}.expert_bias"),
+        routed_experts_op(
+            f"model.layers.{layer_idx}",
+            hf_experts="mlp.experts",
+            tt_experts="mlp.experts",
+            fused=True,
+        ),
     ]
-    state_dict[f"model.layers.{i}.mlp.shared_expert.w2"] = state_dict[
-        f"model.layers.{i}.mlp.shared_experts.down_proj.weight"
-    ]
-    state_dict[f"model.layers.{i}.mlp.shared_expert.w3"] = state_dict[
-        f"model.layers.{i}.mlp.shared_experts.up_proj.weight"
-    ]
-
-    del state_dict[f"model.layers.{i}.mlp.shared_experts.gate_proj.weight"]
-    del state_dict[f"model.layers.{i}.mlp.shared_experts.down_proj.weight"]
-    del state_dict[f"model.layers.{i}.mlp.shared_experts.up_proj.weight"]
-
-    state_dict[f"model.layers.{i}.mlp.expert_bias"] = state_dict[f"model.layers.{i}.mlp.gate.e_score_correction_bias"]
-    del state_dict[f"model.layers.{i}.mlp.gate.e_score_correction_bias"]
+    for wn, hf_proj in GATE_DOWN_UP:
+        # SqueezeLeading is backward-only and must run *after* the rename has
+        # restored the HF key, so it precedes the Rename in the forward list
+        # (apply_tt_to_hf plays the chain in reverse).
+        ops.append(SqueezeLeading(f"{p}.shared_experts.{hf_proj}.weight"))
+        ops.append(Rename(f"{p}.shared_experts.{hf_proj}.weight", f"{p}.shared_expert.{wn}"))
+    ops.append(Drop(f"{p}.tokens_per_expert"))
+    return ops
 
 
-def convert_hf_to_tt_moe(state_dict: dict[str, Tensor]):
-    """Convert MoE weights from HF to TT format in-place."""
-    num_layers = get_max_layer_num(state_dict)
-    for i in range(num_layers):
-        convert_hf_layer_to_tt(state_dict, i)
-
-
-def convert_tt_layer_to_hf(state_dict: dict[str, Tensor], layer_index: int):
-    """Convert a layer from TT to HF format in-place."""
-
-    i = layer_index
-
-    # Load balancing terms
-    if f"model.layers.{i}.mlp.expert_bias" in state_dict:
-        state_dict[f"model.layers.{i}.mlp.gate.e_score_correction_bias"] = state_dict[
-            f"model.layers.{i}.mlp.expert_bias"
-        ]
-        del state_dict[f"model.layers.{i}.mlp.expert_bias"]
-    if f"model.layers.{i}.mlp.tokens_per_expert" in state_dict:
-        del state_dict[f"model.layers.{i}.mlp.tokens_per_expert"]
-
-    # Shared experts
-    if f"model.layers.{i}.mlp.shared_expert.w1" in state_dict:
-        state_dict[f"model.layers.{i}.mlp.shared_experts.gate_proj.weight"] = state_dict[
-            f"model.layers.{i}.mlp.shared_expert.w1"
-        ]
-        state_dict[f"model.layers.{i}.mlp.shared_experts.down_proj.weight"] = state_dict[
-            f"model.layers.{i}.mlp.shared_expert.w2"
-        ]
-        state_dict[f"model.layers.{i}.mlp.shared_experts.up_proj.weight"] = state_dict[
-            f"model.layers.{i}.mlp.shared_expert.w3"
-        ]
-
-        if state_dict[f"model.layers.{i}.mlp.shared_experts.up_proj.weight"].shape[0] == 1:
-            state_dict[f"model.layers.{i}.mlp.shared_experts.up_proj.weight"] = state_dict[
-                f"model.layers.{i}.mlp.shared_experts.up_proj.weight"
-            ][0]
-            state_dict[f"model.layers.{i}.mlp.shared_experts.down_proj.weight"] = state_dict[
-                f"model.layers.{i}.mlp.shared_experts.down_proj.weight"
-            ][0]
-            state_dict[f"model.layers.{i}.mlp.shared_experts.gate_proj.weight"] = state_dict[
-                f"model.layers.{i}.mlp.shared_experts.gate_proj.weight"
-            ][0]
-        del state_dict[f"model.layers.{i}.mlp.shared_expert.w1"]
-        del state_dict[f"model.layers.{i}.mlp.shared_expert.w2"]
-        del state_dict[f"model.layers.{i}.mlp.shared_expert.w3"]
-
-    # Gate / Router
-    if f"model.layers.{i}.mlp.router.gate.weight" in state_dict:
-        state_dict[f"model.layers.{i}.mlp.gate.weight"] = state_dict[f"model.layers.{i}.mlp.router.gate.weight"]
-        del state_dict[f"model.layers.{i}.mlp.router.gate.weight"]
-
-        # Routed experts - convert to per-expert format (compatible with vLLM and transformers)
-        w1 = state_dict.pop(f"model.layers.{i}.mlp.experts.w1")  # (num_experts, moe_dim, dim)
-        w2 = state_dict.pop(f"model.layers.{i}.mlp.experts.w2")  # (num_experts, dim, moe_dim)
-        w3 = state_dict.pop(f"model.layers.{i}.mlp.experts.w3")  # (num_experts, moe_dim, dim)
-
-        num_experts = w1.shape[0]
-        for j in range(num_experts):
-            state_dict[f"model.layers.{i}.mlp.experts.{j}.gate_proj.weight"] = w1[j]
-            state_dict[f"model.layers.{i}.mlp.experts.{j}.down_proj.weight"] = w2[j]
-            state_dict[f"model.layers.{i}.mlp.experts.{j}.up_proj.weight"] = w3[j]
-
-
-def convert_tt_to_hf_moe(state_dict: dict[str, Tensor]):
-    """Convert MoE weights from TT to HF format in-place."""
-    num_layers = get_max_layer_num(state_dict)
-    for i in range(num_layers):
-        convert_tt_layer_to_hf(state_dict, i)
+def conversion_chain(config) -> list[ConvOp]:
+    ops: list[ConvOp] = []
+    for i in range(config.num_hidden_layers):
+        ops.extend(glm_moe_layer_ops(i))
+    return ops
