+++ src/prime_rl/trainer/models/qwen3_moe/converting_qwen3_moe.py
@@ -1,100 +1,19 @@
-import torch
-from torch import Tensor
+"""HF<->prime weight conversion for Qwen3-MoE, as a declarative op chain.
+
+Per layer: router ``mlp.gate.weight`` <-> ``mlp.router.gate.weight`` and the
+routed experts (per-expert gate/down/up <-> stacked w1/w2/w3, with the fused
+transformers-v5 ``gate_up_proj`` input also accepted). No shared experts.
+"""
+
+from __future__ import annotations
+
+from prime_rl.trainer.models.conversion_ops import ConvOp, Rename, routed_experts_op
 
 
-def get_max_layer_num(state_dict: dict[str, Tensor]) -> int:
-    """Get the maximum number of layers in the model."""
-    return max(int(i.split(".")[2]) for i in state_dict.keys() if "model.layers." in i) + 1
-
-
-def convert_hf_layer_to_tt(state_dict: dict[str, Tensor], layer_idx: int):
-    """Convert a layer from HF to TT format in-place."""
-    i = layer_idx
-
-    # Check if this is a MoE layer by looking for gate weight
-    if f"model.layers.{i}.mlp.gate.weight" not in state_dict:
-        return
-
-    state_dict[f"model.layers.{i}.mlp.router.gate.weight"] = state_dict[f"model.layers.{i}.mlp.gate.weight"]
-    del state_dict[f"model.layers.{i}.mlp.gate.weight"]
-
-    # Check if this is the new fused format (transformers 5.0+) or old per-expert format
-    if f"model.layers.{i}.mlp.experts.gate_up_proj" in state_dict:
-        # New fused format: gate_up_proj has shape (num_experts, 2*moe_dim, dim)
-        gate_up_proj = state_dict[f"model.layers.{i}.mlp.experts.gate_up_proj"]
-        down_proj = state_dict[f"model.layers.{i}.mlp.experts.down_proj"]
-
-        num_experts, fused_dim, dim = gate_up_proj.shape
-        moe_dim = fused_dim // 2
-
-        # Split gate_up_proj into w1 (gate) and w3 (up)
-        w1 = gate_up_proj[:, :moe_dim, :]  # Gate: (num_experts, moe_dim, dim)
-        w3 = gate_up_proj[:, moe_dim:, :]  # Up: (num_experts, moe_dim, dim)
-        w2 = down_proj  # Down: (num_experts, dim, moe_dim)
-
-        del state_dict[f"model.layers.{i}.mlp.experts.gate_up_proj"]
-        del state_dict[f"model.layers.{i}.mlp.experts.down_proj"]
-    else:
-        # Old per-expert format
-        num_experts = len([j for j in state_dict.keys() if f"model.layers.{i}.mlp.experts" in j]) // 3
-        if num_experts == 0:
-            return
-
-        dim, moe_dim = state_dict[f"model.layers.{i}.mlp.experts.0.down_proj.weight"].shape
-        w1 = torch.empty(
-            (num_experts, moe_dim, dim), dtype=state_dict[f"model.layers.{i}.mlp.experts.0.down_proj.weight"].dtype
-        )  # Gate
-        w2 = torch.empty(
-            (num_experts, dim, moe_dim), dtype=state_dict[f"model.layers.{i}.mlp.experts.0.down_proj.weight"].dtype
-        )  # Down
-        w3 = torch.empty(
-            (num_experts, moe_dim, dim), dtype=state_dict[f"model.layers.{i}.mlp.experts.0.down_proj.weight"].dtype
-        )  # Up
-        for j in range(num_experts):
-            w1[j].copy_(state_dict[f"model.layers.{i}.mlp.experts.{j}.gate_proj.weight"])
-            w2[j].copy_(state_dict[f"model.layers.{i}.mlp.experts.{j}.down_proj.weight"])
-            w3[j].copy_(state_dict[f"model.layers.{i}.mlp.experts.{j}.up_proj.weight"])
-
-            del state_dict[f"model.layers.{i}.mlp.experts.{j}.gate_proj.weight"]
-            del state_dict[f"model.layers.{i}.mlp.experts.{j}.down_proj.weight"]
-            del state_dict[f"model.layers.{i}.mlp.experts.{j}.up_proj.weight"]
-
-    state_dict[f"model.layers.{i}.mlp.experts.w1"] = w1
-    state_dict[f"model.layers.{i}.mlp.experts.w2"] = w2
-    state_dict[f"model.layers.{i}.mlp.experts.w3"] = w3
-
-
-def convert_tt_layer_to_hf(state_dict: dict[str, Tensor], layer_index: int):
-    """Convert a layer from TT to HF format in-place."""
-    i = layer_index
-    if f"model.layers.{i}.mlp.router.gate.weight" not in state_dict:
-        return
-
-    # Gate / Router
-    state_dict[f"model.layers.{i}.mlp.gate.weight"] = state_dict[f"model.layers.{i}.mlp.router.gate.weight"]
-    del state_dict[f"model.layers.{i}.mlp.router.gate.weight"]
-
-    # Routed experts - convert to per-expert format (compatible with vLLM and transformers)
-    w1 = state_dict.pop(f"model.layers.{i}.mlp.experts.w1")  # (num_experts, moe_dim, dim)
-    w2 = state_dict.pop(f"model.layers.{i}.mlp.experts.w2")  # (num_experts, dim, moe_dim)
-    w3 = state_dict.pop(f"model.layers.{i}.mlp.experts.w3")  # (num_experts, moe_dim, dim)
-
-    num_experts = w1.shape[0]
-    for j in range(num_experts):
-        state_dict[f"model.layers.{i}.mlp.experts.{j}.gate_proj.weight"] = w1[j]
-        state_dict[f"model.layers.{i}.mlp.experts.{j}.down_proj.weight"] = w2[j]
-        state_dict[f"model.layers.{i}.mlp.experts.{j}.up_proj.weight"] = w3[j]
-
-
-def convert_hf_to_tt_moe(state_dict: dict[str, Tensor]):
-    """Convert MoE weights from HF to TT format in-place."""
-    num_layers = get_max_layer_num(state_dict)
-    for i in range(num_layers):
-        convert_hf_layer_to_tt(state_dict, i)
-
-
-def convert_tt_to_hf_moe(state_dict: dict[str, Tensor]):
-    """Convert MoE weights from TT to HF format in-place."""
-    num_layers = get_max_layer_num(state_dict)
-    for i in range(num_layers):
-        convert_tt_layer_to_hf(state_dict, i)
+def conversion_chain(config) -> list[ConvOp]:
+    ops: list[ConvOp] = []
+    for i in range(config.num_hidden_layers):
+        p = f"model.layers.{i}"
+        ops.append(Rename(f"{p}.mlp.gate.weight", f"{p}.mlp.router.gate.weight"))
+        ops.append(routed_experts_op(p, hf_experts="mlp.experts", tt_experts="mlp.experts", fused=True))
+    return ops
