+++ tests/unit/train/models/test_nemotron_h.py
@@ -63,7 +63,7 @@ def get_model_pairs():
     with torch.no_grad():
         state_dict = hf_model.state_dict()
         prime_state_keys = prime_model.state_dict().keys()
-        NemotronHForCausalLM.convert_to_prime(state_dict)
+        prime_model.convert_to_prime(state_dict)
         prime_model.load_state_dict(state_dict)
 
     inject_prime_lm_head(prime_model, chunk_size=None)
@@ -100,56 +100,6 @@ def test_nemotron_h_mamba_moe_only():
     )
     grad_diff = hf_model.model.embeddings.weight.grad - prime_model.model.embed_tokens.weight.grad
     assert torch.allclose(grad_diff, torch.zeros_like(grad_diff), atol=2), f"Max grad diff: {grad_diff.abs().max()}"
-
-
-@pytest.mark.xfail(reason="HF NemotronH now uses fused expert tensors; convert_to_hf produces individual expert format")
-def test_nemotron_h_reverse():
-    """Test reverse: PrimeRL weights loaded into HF model produce identical outputs."""
-    prime_config = NemotronHConfig(
-        **_BASE,
-        layers_block_type=["mamba", "moe", "attention", "moe"],
-        use_grouped_mm=False,
-    )
-    prime_config._attn_implementation = "sdpa"
-
-    hf_config = HFNemotronHConfig(**_BASE, hybrid_override_pattern="ME*E")
-    hf_config._attn_implementation = "sdpa"
-
-    with torch.device("cuda"), default_dtype(torch.float32):
-        prime_model = NemotronHForCausalLM._from_config(prime_config)
-        hf_model = HFNemotronHForCausalLM._from_config(hf_config)
-
-    inject_prime_lm_head(prime_model, chunk_size=None)
-
-    with torch.no_grad():
-        sd = prime_model.state_dict()
-        NemotronHForCausalLM.convert_to_hf(sd)
-        # convert_to_hf produces checkpoint format with "backbone." prefix;
-        # the HF model uses "model." prefix for its state dict
-        keys_to_rename = [k for k in sd if k.startswith("backbone.")]
-        for key in keys_to_rename:
-            sd["model." + key[len("backbone.") :]] = sd.pop(key)
-        hf_model.load_state_dict(sd)
-
-    # Bypass attention to isolate Mamba+MoE matching
-    for layer in hf_model.model.layers:
-        if isinstance(layer.mixer, NemotronHAttention):
-            layer.forward = lambda hidden_states, **kwargs: hidden_states
-    for layer in prime_model.model.layers:
-        if isinstance(layer, NemotronHAttentionLayer):
-            layer.forward = lambda hidden_states, **kwargs: hidden_states
-
-    with torch.device("cuda"), default_dtype(torch.float32):
-        input_ids = torch.randint(0, 256, (1, 32))
-        position_ids = torch.arange(0, 32).unsqueeze(0)
-
-    hf_output = hf_model(input_ids, position_ids=position_ids)
-    prime_output = prime_model(input_ids, position_ids=position_ids)
-
-    logits_diff = prime_output["logits"] - hf_output.logits
-    assert torch.allclose(logits_diff, torch.zeros_like(logits_diff), atol=1e-2), (
-        f"Max logits diff: {logits_diff.abs().max()}"
-    )
 
 
 def test_nemotron_h():
@@ -190,27 +140,6 @@ def test_nemotron_h_backward():
         if p.grad is None or p.grad.norm().item() == 0:
             zero_grads.append(name)
     assert not zero_grads, f"Parameters with zero/no gradients: {zero_grads}"
-
-
-def test_nemotron_h_weight_conversion_roundtrip():
-    """Verify PrimeRL -> HF -> PrimeRL conversion preserves all weights."""
-    prime_config = NemotronHConfig(
-        **_BASE,
-        layers_block_type=["mamba", "moe", "attention", "moe"],
-        use_grouped_mm=False,
-    )
-    model = NemotronHForCausalLM(prime_config).to("cuda")
-    original_sd = {k: v.clone() for k, v in model.state_dict().items()}
-
-    sd = model.state_dict()
-    NemotronHForCausalLM.convert_to_hf(sd)
-    assert NemotronHForCausalLM.is_hf_state_dict(sd)
-    NemotronHForCausalLM.convert_to_prime(sd)
-    assert NemotronHForCausalLM.is_prime_state_dict(sd)
-
-    for key in original_sd:
-        assert key in sd, f"Missing key after roundtrip: {key}"
-        assert torch.equal(original_sd[key], sd[key]), f"Value mismatch for {key}"
 
 
 def test_nemotron_h_hybrid_override_pattern():
